Introduce IXXAT_DEBUG in replacement to DEBUG
trace_printk() SHOULDN'T be used outside of a DEBUG kernel but nothing
prevents to build the OOT driver with -DDEBUG.

Using IXXAT_DEBUG (instead of DEBUG) allows to get simple traces in
dmesg log. Of course, in a DEBUG kernel, trace_printk() will be used.

Code is rearranged to move these local definitions on top of the C source
files.